Splits up Server.cs
