+++ TGServiceInterface/Server.cs
@@ -145,64 +145,4 @@ namespace TGServiceInterface
 			}
 		}
 	}
-
-	/// <summary>
-	/// Used for testing connections to the service without authentication
-	/// </summary>
-	[ServiceContract]
-	public interface ITGConnectivity
-	{
-		/// <summary>
-		/// Does nothing on the server end, but if the call completes, you can be sure you are connected. WCF won't throw until you try until you actually use the API
-		/// </summary>
-		[OperationContract]
-		void VerifyConnection();
-	}
-
-	/// <summary>
-	/// Manage the group that is used to access the service, can only be used by an administrator
-	/// </summary>
-	[ServiceContract]
-	public interface ITGAdministration
-	{
-		/// <summary>
-		/// Returns the name of the windows group allowed to use the service other than administrator
-		/// </summary>
-		/// <returns>The name of the windows group allowed to use the service other than administrator, "ADMIN" if it's unset, null on failure</returns>
-		[OperationContract]
-		string GetCurrentAuthorizedGroup();
-
-		/// <summary>
-		/// Searches the windows machine for the group named <paramref name="groupName"/>, sets it as the authorized group if it's found
-		/// </summary>
-		/// <param name="groupName">The name of the group to search for or null to clear the setting</param>
-		/// <returns>The full name of the group that is now authorized on success, null on failure, "ADMIN" on clearing</returns>
-		[OperationContract]
-		string SetAuthorizedGroup(string groupName);
-	}
-
-	/// <summary>
-	/// Interface for managing the service
-	/// </summary>
-	[ServiceContract]
-	public interface ITGSService
-	{
-
-		/// <summary>
-		/// Next stop of the service will not close DD and sets a flag for it to reattach once it restarts
-		/// </summary>
-		[OperationContract]
-		void PrepareForUpdate();
-
-		[OperationContract]
-		[Obsolete("StopForUpdate is deprecated, please use PrepareForUpdate instead")]
-		void StopForUpdate();
-
-		/// <summary>
-		/// Retrieve's the service's version
-		/// </summary>
-		/// <returns>The service's version</returns>
-		[OperationContract]
-		string Version();
-	}
 }
